+++ apps/web/src/components/toolbox/filter-tool/filter-tool-by-path.tsx
@@ -1,12 +1,8 @@
-import { CollapsibleTrigger } from '@/components/ui/collapsible';
-import { CollapsibleContent } from '@/components/ui/collapsible';
-import { Collapsible } from '@/components/ui/collapsible';
 import { ScrollArea } from '@/components/ui/scroll-area';
 import { GraphManager } from '@/lib/graph-manager';
 import { getLeafIds } from '@/lib/tree-node/get-leaf-ids';
 import { createTreeNodeFromGraph } from '@ts2uml/core/src/tree-node/create-tree-node-from-graph.ts';
 import { type TreeNode, createMsgUpdateVisibleNodes } from '@ts2uml/models';
-import { ChevronRight } from 'lucide-react';
 import { useEffect, useState } from 'react';
 import { FilterToolByPathTreeNode } from './filter-tool-by-path-tree-node';
 
@@ -37,19 +33,13 @@ export function FilterToolByPath() {
   function handleFilterByPathChange(treeNode: TreeNode) {
     const currentFilteredNodeIds = gm.getGraph().config.nodes.filter.filter_node;
     let newFilteredNodeIds = currentFilteredNodeIds;
-    let nodeIdsToAdd: string[] = [];
-    let nodeIdsToRemove: string[] = [];
 
     if (treeNode.isElement) {
       const res = getNewFilteredNodesFromElementNodeToggle(treeNode, currentFilteredNodeIds);
       newFilteredNodeIds = res.newFilteredNodeIds;
-      nodeIdsToAdd = res.nodeIdsToAdd;
-      nodeIdsToRemove = res.nodeIdsToRemove;
     } else {
       const res = getNewFilteredNodesFromFolderOrFileNodeToggle(treeNode, currentFilteredNodeIds);
       newFilteredNodeIds = res.newFilteredNodeIds;
-      nodeIdsToAdd = res.nodeIdsToAdd;
-      nodeIdsToRemove = res.nodeIdsToRemove;
     }
 
     // Update the graph
@@ -64,7 +54,7 @@ export function FilterToolByPath() {
     });
 
     updateTree();
-    window.postMessage(createMsgUpdateVisibleNodes({ nodeIdsToAdd, nodeIdsToRemove }));
+    window.postMessage(createMsgUpdateVisibleNodes());
   }
 
   function getNewFilteredNodesFromElementNodeToggle(treeNode: TreeNode, originalFilteredNodes: string[]) {
@@ -108,22 +98,13 @@ export function FilterToolByPath() {
   }
 
   return (
-    <Collapsible defaultOpen={true} className="flex flex-col items-start justify-start">
-      <div className="flex w-full items-center justify-between">
-        <span className="text-xs">{'By folder / file / node'}</span>
-        <CollapsibleTrigger asChild>
-          <div className="chevron-right-collapsible-icon">
-            <ChevronRight className="chevron-icon h-4 w-4" />
-          </div>
-        </CollapsibleTrigger>
-      </div>
-      <CollapsibleContent>
-        <ScrollArea className="h-80 max-h-80 pr-4">
-          {sortedTreeNodes.map((node) => (
-            <FilterToolByPathTreeNode key={node.id} tree={node} onClick={handleFilterByPathChange} />
-          ))}
-        </ScrollArea>
-      </CollapsibleContent>
-    </Collapsible>
+    <div className="flex flex-col items-start justify-start gap-2">
+      <span className="text-xs">{'By folder / file / node'}</span>
+      <ScrollArea className="h-80 max-h-80 pr-4">
+        {sortedTreeNodes.map((node) => (
+          <FilterToolByPathTreeNode key={node.id} tree={node} onClick={handleFilterByPathChange} />
+        ))}
+      </ScrollArea>
+    </div>
   );
 }