+++ apps/web/src/components/toolbox/filter-tool/filter-tool-by-type.tsx
@@ -1,7 +1,4 @@
 import { CheckboxPartial } from '@/components/checkbox-partial';
-import { CollapsibleContent } from '@/components/ui/collapsible';
-import { CollapsibleTrigger } from '@/components/ui/collapsible';
-import { Collapsible } from '@/components/ui/collapsible';
 import { GraphManager } from '@/lib/graph-manager';
 import {
   type CheckboxPartialCheckedStatus,
@@ -9,7 +6,6 @@ import {
   ZNodeType,
   createMsgUpdateVisibleNodes,
 } from '@ts2uml/models';
-import { ChevronRight } from 'lucide-react';
 import { useState } from 'react';
 
 export function FilterToolByType() {
@@ -38,13 +34,9 @@ export function FilterToolByType() {
 
   function handleFilterTypeChange(type: NodeType) {
     let filterByType = gm.getGraph().config.nodes.filter.filter_type;
-    let nodeIdsToAdd: string[] = [];
-    let nodeIdsToRemove: string[] = [];
     if (filterByType.includes(type)) {
-      nodeIdsToAdd = getNodeIdsByType(type);
       filterByType = filterByType.filter((t) => t !== type);
     } else {
-      nodeIdsToRemove = getNodeIdsByType(type);
       filterByType.push(type);
     }
 
@@ -59,37 +51,20 @@ export function FilterToolByType() {
     });
 
     updateCheckedStatusByType();
-    window.postMessage(createMsgUpdateVisibleNodes({ nodeIdsToAdd, nodeIdsToRemove }));
-  }
-
-  function getNodeIdsByType(type: NodeType) {
-    const nodeIds: string[] = [];
-    for (const node of gm.getGraph().nodes) {
-      if (node.type === type) {
-        nodeIds.push(node.id);
-      }
-    }
-    return nodeIds;
+    window.postMessage(createMsgUpdateVisibleNodes());
   }
 
   return (
-    <Collapsible defaultOpen={true} className="flex flex-col items-start justify-start">
-      <div className="flex w-full items-center justify-between">
-        <span className="pb-1 text-xs">By type</span>
-        <CollapsibleTrigger asChild>
-          <div className="chevron-right-collapsible-icon">
-            <ChevronRight className="chevron-icon h-4 w-4" />
-          </div>
-        </CollapsibleTrigger>
-      </div>
-      <CollapsibleContent className="">
+    <div className="flex flex-col items-start justify-start gap-2">
+      <span className="text-xs">By type</span>
+      <div className="">
         {ZNodeType.options.map((t) => (
           <div key={t} className="flex items-center justify-start gap-1">
             <CheckboxPartial checked={checkedStatusByType[t]} onClick={() => handleFilterTypeChange(t)} />
             <span className="text-sm">{t}</span>
           </div>
         ))}
-      </CollapsibleContent>
-    </Collapsible>
+      </div>
+    </div>
   );
 }