+++ apps/web/src/components/toolbox/filter-tool/filter-tool.tsx
@@ -7,6 +7,7 @@ import { Filter } from 'lucide-react';
 import { useState } from 'react';
 import { Button } from '../../ui/button';
 import { Popover, PopoverContent, PopoverTrigger } from '../../ui/popover';
+import { FilterToolByName } from './filter-tool-by-name';
 import { FilterToolByPath } from './filter-tool-by-path';
 import { FilterToolByType } from './filter-tool-by-type';
 
@@ -29,17 +30,20 @@ export function FilterTool() {
           </TooltipContent>
         </Tooltip>
       </TooltipProvider>
-      <PopoverContent className="flex min-h-80 min-w-64 flex-col border border-primary" sideOffset={10}>
+      <PopoverContent autoFocus={false} className="flex min-h-80 flex-col border border-primary" sideOffset={10}>
         <div className="flex flex-col">
           <div className="flex items-center justify-between">
             <span className="font-medium">Filter nodes</span>
             <Filter className="h-3 w-3" />
           </div>
-          <Separator className="mt-1 mb-2 bg-foreground/10" orientation="horizontal" />
-
-          <FilterToolByType />
-          <Separator className="mt-1 mb-2 bg-border/20" orientation="horizontal" />
-          <FilterToolByPath />
+          <Separator className="mt-1 mb-2 bg-foreground/20" orientation="horizontal" />
+          <div className="flex gap-4">
+            <FilterToolByName />
+            <Separator className="h-80 bg-foreground/10" orientation="vertical" />
+            <FilterToolByPath />
+            <Separator className="h-80 bg-foreground/10" orientation="vertical" />
+            <FilterToolByType />
+          </div>
         </div>
       </PopoverContent>
     </Popover>