+++ apps/web/src/components/toolbox/layout-tool/layout-tool-layout-algorithm.tsx
@@ -1,11 +1,7 @@
-import { CollapsibleContent } from '@/components/ui/collapsible';
-import { CollapsibleTrigger } from '@/components/ui/collapsible';
-import { Collapsible } from '@/components/ui/collapsible';
 import { RadioGroupItem } from '@/components/ui/radio-group';
 import { RadioGroup } from '@/components/ui/radio-group';
 import { GraphManager } from '@/lib/graph-manager';
 import { type LayoutAlgorithm, ZLayoutAlgorithm, createMsgUpdateLayoutAlgorithm } from '@ts2uml/models';
-import { ChevronRight } from 'lucide-react';
 import { useState } from 'react';
 
 export function LayoutToolLayoutAlgorithm() {
@@ -25,16 +21,9 @@ export function LayoutToolLayoutAlgorithm() {
   };
 
   return (
-    <Collapsible defaultOpen={true} className="flex flex-col items-start justify-start">
-      <div className="flex w-full items-center justify-between">
-        <span className="pb-1 text-xs">Layout algorithm</span>
-        <CollapsibleTrigger asChild>
-          <div className="chevron-right-collapsible-icon">
-            <ChevronRight className="chevron-icon h-4 w-4" />
-          </div>
-        </CollapsibleTrigger>
-      </div>
-      <CollapsibleContent className="">
+    <div className="flex flex-col items-start justify-start gap-2">
+      <span className="text-xs">Layout algorithm</span>
+      <div className="flex flex-col items-start justify-start">
         <RadioGroup value={layoutAlgorithm} onValueChange={handleLayoutAlgorithmChange}>
           {ZLayoutAlgorithm.options.map((t) => (
             <div className="flex items-center justify-start gap-2" key={t}>
@@ -43,7 +32,7 @@ export function LayoutToolLayoutAlgorithm() {
             </div>
           ))}
         </RadioGroup>
-      </CollapsibleContent>
-    </Collapsible>
+      </div>
+    </div>
   );
 }