+++ apps/web/src/components/toolbox/layout-tool/layout-tool-link-path-algorithm.tsx
@@ -1,11 +1,8 @@
-import { CollapsibleContent } from '@/components/ui/collapsible';
-import { CollapsibleTrigger } from '@/components/ui/collapsible';
-import { Collapsible } from '@/components/ui/collapsible';
 import { RadioGroupItem } from '@/components/ui/radio-group';
 import { RadioGroup } from '@/components/ui/radio-group';
 import { GraphManager } from '@/lib/graph-manager';
 import { type LinkPathAlgorithm, ZLinkPathAlgorithm, createMsgUpdateLinkPathAlgorithm } from '@ts2uml/models';
-import { ChevronRight } from 'lucide-react';
+import { ChevronLeft, Slash, Spline } from 'lucide-react';
 import { useState } from 'react';
 
 export function LayoutToolLinkPathAlgorithm() {
@@ -29,25 +26,21 @@ export function LayoutToolLinkPathAlgorithm() {
   };
 
   return (
-    <Collapsible defaultOpen={true} className="flex flex-col items-start justify-start">
-      <div className="flex w-full items-center justify-between">
-        <span className="pb-1 text-xs">Link path algorithm</span>
-        <CollapsibleTrigger asChild>
-          <div className="chevron-right-collapsible-icon">
-            <ChevronRight className="chevron-icon h-4 w-4" />
-          </div>
-        </CollapsibleTrigger>
-      </div>
-      <CollapsibleContent className="">
+    <div className="flex flex-col items-start justify-start gap-2">
+      <span className="text-xs">Link path algorithm</span>
+      <div className="flex flex-col items-start justify-start">
         <RadioGroup value={linkPathAlgorithm} onValueChange={handleLinkPathAlgorithmChange}>
           {ZLinkPathAlgorithm.options.map((t) => (
             <div className="flex items-center justify-start gap-2" key={t}>
               <RadioGroupItem value={t} id={t} />
+              {t === 'straight' && <Slash className="h-4 w-4" />}
+              {t === 'bezier' && <Spline className="h-4 w-4" />}
+              {t === 'step' && <ChevronLeft className="h-4 w-4 rotate-45" />}
               <span className="text-sm">{t}</span>
             </div>
           ))}
         </RadioGroup>
-      </CollapsibleContent>
-    </Collapsible>
+      </div>
+    </div>
   );
 }