+++ packages/models/src/types/config/config-nodes-filter.ts
@@ -1,10 +1,20 @@
 import { z } from 'zod';
 import { type NodeType, ZNodeType } from '../graph/node-type.ts';
+import {
+  type ConfigNodesFilterName,
+  ZConfigNodesFilterName,
+  createConfigNodesFilterName,
+} from './config-nodes-filter-name.ts';
 
 /**
  * Interface defining filters for nodes to be displayed in the diagram.
  */
 export interface ConfigNodesFilter {
+  /**
+   * Filter nodes by their name.
+   */
+  filter_name: ConfigNodesFilterName;
+
   /**
    * Array of strings to filter nodes by their file path or name.
    * This array can contain directories, files and concrete node ids.
@@ -25,12 +35,14 @@ export interface ConfigNodesFilter {
 }
 
 export const ZConfigNodesFilter = z.object({
+  filter_name: ZConfigNodesFilterName,
   filter_node: z.array(z.string()).optional(),
   filter_type: ZNodeType.array().optional(),
 }) as z.ZodType<ConfigNodesFilter>;
 
 export function createConfigNodesFilter(data?: Partial<ConfigNodesFilter>): ConfigNodesFilter {
   return ZConfigNodesFilter.parse({
+    filter_name: createConfigNodesFilterName(data?.filter_name),
     filter_node: data?.filter_node ?? [],
     filter_type: data?.filter_type ?? [],
   });